+++ ios/Plugin/Models/Inputs/DeleteEventInput.swift
@@ -2,10 +2,12 @@ import Capacitor
 import EventKit
 
 struct DeleteEventInput {
+    private let commit: Bool
     private let id: String
     private let span: EKSpan
 
     init(call: CAPPluginCall) throws {
+        self.commit = call.getBool("commit", true)
         guard let id = call.getString("id") else {
             throw PluginError.idMissing
         }
@@ -15,6 +17,10 @@ struct DeleteEventInput {
         } else {
             self.span = .thisEvent
         }
+    }
+
+    func getCommit() -> Bool {
+        return commit
     }
 
     func getId() -> String {