+++ ios/Plugin/Models/Inputs/DeleteEventsByIdInput.swift
@@ -2,10 +2,12 @@ import Capacitor
 import EventKit
 
 struct DeleteEventsByIdInput {
+    private let commit: Bool
     private let ids: [String]
     private let span: EKSpan
 
     init(call: CAPPluginCall) throws {
+        self.commit = call.getBool("commit", true)
         guard let ids = call.getArray("ids") as? [String] else {
             throw PluginError.idMissing
         }
@@ -15,6 +17,10 @@ struct DeleteEventsByIdInput {
         } else {
             self.span = .thisEvent
         }
+    }
+
+    func getCommit() -> Bool {
+        return commit
     }
 
     func getIds() -> [String] {