+++ ios/Plugin/Utils/ImplementationHelper.swift
@@ -133,11 +133,11 @@ struct ImplementationHelper {
         try eventStore.remove(reminder, commit: true)
     }
 
-    static func deleteEvent(_ id: String, _ span: EKSpan, _ eventStore: EKEventStore) throws {
+    static func deleteEvent(commit: Bool, _ id: String, _ span: EKSpan, _ eventStore: EKEventStore) throws {
         guard let event = eventStore.event(withIdentifier: id) else {
             throw PluginError.eventNotFound
         }
-        try eventStore.remove(event, span: span, commit: true)
+        try eventStore.remove(event, span: span, commit: commit)
     }
 
     static func calendarsSetToJSArray(_ calendars: Set<EKCalendar>) -> [JSObject] {